feat(controller): add CommandController interface, will extend from TouchScreen
The CommandController interface extends the TouchScreen interface and adds a default method select() that receives the keyboard() method from the TouchScreen interface and a Map that will receive the Number from 1 to 4 typed by the user and the CommandController interface itself as a value to act as Command FunctionalInterface on invocation. All implementations of the CommandController interface will have the Map in their constructors and call the boolean select() method at the end of their void execute() methods.